SPSH-1026 (#227)
* Array of PersonIds instead of all of them

* Refactored getAllUebersichte into PersonStore instead of Personenkontext

* lint

* Migrated getPersonWithUebersicht

* fixed tests

* Fixed build

* Fixed build

* If no personIds then just return

* null for backend

* return and avoid another request if no personIds

* Changed query to Post with  Body
