+++ src/stores/PersonStore.spec.ts
@@ -1,7 +1,11 @@
-import type {
-  PersonFrontendControllerFindPersons200Response,
-  PersonendatensatzResponse,
-  TokenStateResponse,
+import {
+  OrganisationsTyp,
+  RollenMerkmal,
+  type DBiamPersonenuebersichtControllerFindPersonenuebersichten200Response,
+  type DBiamPersonenuebersichtResponse,
+  type PersonFrontendControllerFindPersons200Response,
+  type PersonendatensatzResponse,
+  type TokenStateResponse,
 } from '@/api-client/generated';
 import { usePersonStore, type PersonStore, type Personendatensatz } from './PersonStore';
 import ApiService from '@/services/ApiService';
@@ -20,13 +24,13 @@ describe('PersonStore', () => {
   });
 
   it('should initalize state correctly', () => {
-    expect(personStore.allPersons).toEqual([]);
     expect(personStore.errorCode).toEqual('');
     expect(personStore.loading).toBe(false);
   });
 
   describe('getAllPersons', () => {
-    it('should load Persons and update state', async () => {
+    it('should load persons and their overviews, and update state', async () => {
+      // Mock data for persons
       const mockPersons: PersonendatensatzResponse[] = [
         {
           person: {
@@ -48,66 +52,135 @@ describe('PersonStore', () => {
         },
       ] as PersonendatensatzResponse[];
 
-      const mockResponse: PersonFrontendControllerFindPersons200Response = {
+      // Mock response for persons
+      const mockPersonsResponse: PersonFrontendControllerFindPersons200Response = {
         offset: 0,
         limit: 2,
         total: 2,
         items: mockPersons,
       };
-      mockadapter.onGet('/api/personen-frontend').replyOnce(200, mockResponse, {});
-      const getAllPersonPromise: Promise<void> = personStore.getAllPersons({});
+
+      // Mock data for person overviews
+      const mockUebersichten: DBiamPersonenuebersichtControllerFindPersonenuebersichten200Response = {
+        total: 2,
+        offset: 0,
+        limit: 2,
+        items: [
+          {
+            personId: '1234',
+            vorname: 'Samuel',
+            nachname: 'Vimes',
+            benutzername: 'string',
+            lastModifiedZuordnungen: '08.02.2024',
+            zuordnungen: [
+              {
+                sskId: 'string',
+                rolleId: 'string',
+                sskName: 'string',
+                sskDstNr: 'string',
+                rolle: 'string',
+                typ: OrganisationsTyp.Klasse,
+                administriertVon: 'string',
+                editable: true,
+                merkmale: [] as unknown as RollenMerkmal,
+              },
+            ],
+          },
+        ],
+      };
+
+      mockadapter.onGet('/api/personen-frontend').replyOnce(200, mockPersonsResponse);
+
+      // Update the mock POST request with the appropriate body
+      const personIds = mockPersons.map((person) => person.person.id);
+      mockadapter.onPost('/api/dbiam/personenuebersicht', { personIds }).replyOnce(200, mockUebersichten);
+
+      const getAllPersonsPromise: Promise<void> = personStore.getAllPersons({});
       expect(personStore.loading).toBe(true);
-      await getAllPersonPromise;
-      expect(personStore.allPersons).toEqual([...mockPersons]);
+      await getAllPersonsPromise;
       expect(personStore.loading).toBe(false);
     });
 
-    it('should load persons according to filter', async () => {
-      const mockPersonsWithFilter: PersonendatensatzResponse[] = [
+    it('should handle string error response', async () => {
+      const mockPersons: PersonendatensatzResponse[] = [
         {
           person: {
-            id: '123456',
+            id: '1234',
             name: {
               familienname: 'Vimes',
-              vorname: 'Susan',
+              vorname: 'Samuel',
+            },
+          },
+        },
+        {
+          person: {
+            id: '5678',
+            name: {
+              familienname: 'von Lipwig',
+              vorname: 'Moist',
             },
           },
         },
       ] as PersonendatensatzResponse[];
 
-      const mockResponseWithFilter: PersonFrontendControllerFindPersons200Response = {
+      const mockPersonsResponse: PersonFrontendControllerFindPersons200Response = {
         offset: 0,
-        limit: 1,
-        total: 1,
-        items: mockPersonsWithFilter,
+        limit: 2,
+        total: 2,
+        items: mockPersons,
       };
 
-      mockadapter.onGet('/api/personen-frontend?suchFilter=Sus').replyOnce(200, mockResponseWithFilter, {});
-      const getAllPersonPromise: Promise<void> = personStore.getAllPersons({ searchFilter: 'Sus' });
-      expect(personStore.loading).toBe(true);
-      await getAllPersonPromise;
-      expect(personStore.allPersons).toEqual([...mockPersonsWithFilter]);
-      expect(personStore.loading).toBe(false);
-    });
+      mockadapter.onGet('/api/personen-frontend').replyOnce(200, mockPersonsResponse);
 
-    it('should handle string error', async () => {
-      mockadapter.onGet('/api/personen-frontend').replyOnce(500, 'some mock server error');
-      const getAllPersonPromise: Promise<void> = personStore.getAllPersons({});
+      const personIds: string[] = mockPersons.map((person) => person.person.id);
+      mockadapter.onPost('/api/dbiam/personenuebersicht', { personIds }).replyOnce(500, 'Some error occurred');
+
+      const getAllPersonsPromise: Promise<void> = personStore.getAllPersons({});
       expect(personStore.loading).toBe(true);
-      await getAllPersonPromise;
-      expect(personStore.allPersons).toEqual([]);
+      await getAllPersonsPromise;
+      expect(personStore.loading).toBe(false);
       expect(personStore.errorCode).toEqual('UNSPECIFIED_ERROR');
-      expect(personStore.loading).toBe(false);
     });
 
-    it('should handle error code', async () => {
-      mockadapter.onGet('/api/personen-frontend').replyOnce(500, { code: 'some mock server error' });
-      const getAllPersonPromise: Promise<void> = personStore.getAllPersons({});
+    it('should handle error code in response', async () => {
+      const mockPersons: PersonendatensatzResponse[] = [
+        {
+          person: {
+            id: '1234',
+            name: {
+              familienname: 'Vimes',
+              vorname: 'Samuel',
+            },
+          },
+        },
+        {
+          person: {
+            id: '5678',
+            name: {
+              familienname: 'von Lipwig',
+              vorname: 'Moist',
+            },
+          },
+        },
+      ] as PersonendatensatzResponse[];
+
+      const mockPersonsResponse: PersonFrontendControllerFindPersons200Response = {
+        offset: 0,
+        limit: 2,
+        total: 2,
+        items: mockPersons,
+      };
+
+      mockadapter.onGet('/api/personen-frontend').replyOnce(200, mockPersonsResponse);
+
+      const personIds: string[] = mockPersons.map((person) => person.person.id);
+      mockadapter.onPost('/api/dbiam/personenuebersicht', { personIds }).replyOnce(500, { code: 'SERVER_ERROR' });
+
+      const getAllPersonsPromise: Promise<void> = personStore.getAllPersons({});
       expect(personStore.loading).toBe(true);
-      await getAllPersonPromise;
-      expect(personStore.allPersons).toEqual([]);
-      expect(personStore.errorCode).toEqual('some mock server error');
+      await getAllPersonsPromise;
       expect(personStore.loading).toBe(false);
+      expect(personStore.errorCode).toEqual('SERVER_ERROR');
     });
   });
 
@@ -294,6 +367,56 @@ describe('PersonStore', () => {
       const get2FASoftwareQRCodePromise: Promise<void> = personStore.get2FASoftwareQRCode(personId);
       expect(personStore.loading).toBe(true);
       await rejects(get2FASoftwareQRCodePromise);
+      expect(personStore.errorCode).toEqual('some mock server error');
+      expect(personStore.loading).toBe(false);
+    });
+  });
+
+  describe('getPersonenuebersichtById', () => {
+    it('should get Personenuebersicht', async () => {
+      const mockResponse: DBiamPersonenuebersichtResponse = {
+        personId: '1',
+        vorname: 'string',
+        nachname: 'string',
+        benutzername: 'string',
+        lastModifiedZuordnungen: '08.02.2024',
+        zuordnungen: [
+          {
+            sskId: 'string',
+            rolleId: 'string',
+            sskName: 'string',
+            sskDstNr: 'string',
+            rolle: 'string',
+            typ: OrganisationsTyp.Klasse,
+            administriertVon: 'string',
+            editable: true,
+            merkmale: [] as unknown as RollenMerkmal,
+          },
+        ],
+      };
+
+      mockadapter.onGet('/api/dbiam/personenuebersicht/1').replyOnce(200, mockResponse);
+      const getPersonenuebersichtByIdPromise: Promise<void> = personStore.getPersonenuebersichtById('1');
+      expect(personStore.loading).toBe(true);
+      await getPersonenuebersichtByIdPromise;
+      expect(personStore.personenuebersicht).toEqual(mockResponse);
+      expect(personStore.loading).toBe(false);
+    });
+
+    it('should handle string error', async () => {
+      mockadapter.onGet('/api/dbiam/personenuebersicht/1').replyOnce(500, 'some error');
+      const getPersonenuebersichtByIdPromise: Promise<void> = personStore.getPersonenuebersichtById('1');
+      expect(personStore.loading).toBe(true);
+      await getPersonenuebersichtByIdPromise;
+      expect(personStore.errorCode).toEqual('UNSPECIFIED_ERROR');
+      expect(personStore.loading).toBe(false);
+    });
+
+    it('should handle error code', async () => {
+      mockadapter.onGet('/api/dbiam/personenuebersicht/1').replyOnce(500, { code: 'some mock server error' });
+      const getPersonenuebersichtByIdPromise: Promise<void> = personStore.getPersonenuebersichtById('1');
+      expect(personStore.loading).toBe(true);
+      await getPersonenuebersichtByIdPromise;
       expect(personStore.errorCode).toEqual('some mock server error');
       expect(personStore.loading).toBe(false);
     });
