+++ src/stores/PersonStore.ts
@@ -2,22 +2,34 @@ import { defineStore, type Store, type StoreDefinition } from 'pinia';
 import { isAxiosError, type AxiosResponse } from 'axios';
 import {
   Class2FAApiFactory,
+  DbiamPersonenuebersichtApiFactory,
+  OrganisationsTyp,
   PersonenApiFactory,
   PersonenFrontendApiFactory,
+  RollenMerkmal,
   type Class2FAApiInterface,
   type DbiamCreatePersonWithContextBodyParams,
+  type DbiamPersonenuebersichtApiInterface,
+  type DBiamPersonenuebersichtControllerFindPersonenuebersichten200Response,
+  type DBiamPersonenuebersichtResponse,
   type PersonenApiInterface,
   type PersonendatensatzResponse,
   type PersonenFrontendApiInterface,
+  type PersonenuebersichtBodyParams,
   type PersonFrontendControllerFindPersons200Response,
   type TokenInitBodyParams,
   type TokenStateResponse,
 } from '../api-client/generated/api';
 import axiosApiInstance from '@/services/ApiService';
-import { type DbiamPersonenkontextBodyParams } from './PersonenkontextStore';
+import { type DbiamPersonenkontextBodyParams, type Zuordnung } from './PersonenkontextStore';
 
 const personenApi: PersonenApiInterface = PersonenApiFactory(undefined, '', axiosApiInstance);
 const personenFrontendApi: PersonenFrontendApiInterface = PersonenFrontendApiFactory(undefined, '', axiosApiInstance);
+const personenuebersichtApi: DbiamPersonenuebersichtApiInterface = DbiamPersonenuebersichtApiFactory(
+  undefined,
+  '',
+  axiosApiInstance,
+);
 
 const twoFactorApi: Class2FAApiInterface = Class2FAApiFactory(undefined, '', axiosApiInstance);
 
@@ -30,6 +42,35 @@ export type Person = {
   referrer: string | null;
   personalnummer?: string | null;
 };
+
+type PersonenWithRolleAndZuordnung = {
+  rollen: string;
+  administrationsebenen: string;
+  klassen: string;
+  person: Person;
+}[];
+
+export type PersonWithUebersicht =
+  | {
+      personId: string;
+      vorname: string;
+      nachname: string;
+      benutzername: string;
+      lastModifiedZuordnungen: string | null;
+      zuordnungen: {
+        klasse?: string | undefined;
+        sskId: string;
+        rolleId: string;
+        sskName: string;
+        sskDstNr: string;
+        rolle: string;
+        administriertVon: string;
+        typ: OrganisationsTyp;
+        editable: boolean;
+        merkmale: RollenMerkmal;
+      }[];
+    }
+  | undefined;
 
 export type PersonTableItem = {
   person: Person;
@@ -54,12 +95,13 @@ export type TwoFactorState = {
 export type { PersonendatensatzResponse };
 
 type PersonState = {
-  allPersons: Array<Personendatensatz>;
   errorCode: string;
   loading: boolean;
   totalPersons: number;
   currentPerson: Personendatensatz | null;
   twoFactorState: TwoFactorState;
+  personenWithUebersicht: PersonenWithRolleAndZuordnung | null;
+  personenuebersicht: DBiamPersonenuebersichtResponse | null;
 };
 
 export type PersonFilter = {
@@ -79,6 +121,7 @@ type PersonActions = {
   deletePersonById: (personId: string) => Promise<void>;
   get2FAState: (personId: string) => Promise<void>;
   get2FASoftwareQRCode: (personId: string) => Promise<void>;
+  getPersonenuebersichtById: (personId: string) => Promise<void>;
 };
 
 export type PersonStore = Store<'personStore', PersonState, PersonGetters, PersonActions>;
@@ -87,7 +130,8 @@ export const usePersonStore: StoreDefinition<'personStore', PersonState, PersonG
   id: 'personStore',
   state: (): PersonState => {
     return {
-      allPersons: [],
+      personenWithUebersicht: null,
+      personenuebersicht: null,
       errorCode: '',
       loading: false,
       totalPersons: 0,
@@ -107,6 +151,7 @@ export const usePersonStore: StoreDefinition<'personStore', PersonState, PersonG
     async getAllPersons(filter: PersonFilter) {
       this.loading = true;
       try {
+        // Fetch all persons
         const { data }: AxiosResponse<PersonFrontendControllerFindPersons200Response> =
           await personenFrontendApi.personFrontendControllerFindPersons(
             filter.offset,
@@ -120,8 +165,62 @@ export const usePersonStore: StoreDefinition<'personStore', PersonState, PersonG
             filter.searchFilter,
           );
 
-        this.allPersons = data.items;
+        // Store the fetched persons
+        const allPersons: PersonendatensatzResponse[] = data.items;
         this.totalPersons = +data.total;
+
+        // Fetch overviews for all persons
+        const personIds: string[] = data.items.map((person: PersonendatensatzResponse) => person.person.id);
+        if (personIds.length === 0) {
+          this.personenWithUebersicht = null;
+          return;
+        }
+        const bodyParams: PersonenuebersichtBodyParams = {
+          personIds: personIds
+        }
+        const { data: uebersichten }: { data: DBiamPersonenuebersichtControllerFindPersonenuebersichten200Response } =
+
+          await personenuebersichtApi.dBiamPersonenuebersichtControllerFindPersonenuebersichten(bodyParams);
+        const allUebersichten: DBiamPersonenuebersichtControllerFindPersonenuebersichten200Response = uebersichten;
+
+        // Aggregate the personen with their uebersichten
+        this.personenWithUebersicht = allPersons.map((person: Personendatensatz) => {
+          const uebersicht: PersonWithUebersicht = allUebersichten.items.find(
+            (ueb: PersonWithUebersicht) => ueb?.personId === person.person.id,
+          );
+
+          const uniqueRollen: Set<string> = new Set<string>();
+          uebersicht?.zuordnungen.forEach((zuordnung: Zuordnung) => uniqueRollen.add(zuordnung.rolle));
+          const rollenZuordnungen: string = uniqueRollen.size > 0 ? Array.from(uniqueRollen).join(', ') : '---';
+
+          const uniqueAdministrationsebenen: Set<string> = new Set<string>();
+          uebersicht?.zuordnungen
+            .filter((zuordnung: Zuordnung) => zuordnung.typ !== OrganisationsTyp.Klasse)
+            .forEach((zuordnung: Zuordnung) =>
+              uniqueAdministrationsebenen.add(zuordnung.sskDstNr ? zuordnung.sskDstNr : zuordnung.sskName),
+            );
+          const administrationsebenen: string =
+            uniqueAdministrationsebenen.size > 0 ? Array.from(uniqueAdministrationsebenen).join(', ') : '---';
+
+          const klassenZuordnungen: string = uebersicht?.zuordnungen.some(
+            (zuordnung: Zuordnung) => zuordnung.typ === OrganisationsTyp.Klasse,
+          )
+            ? uebersicht.zuordnungen
+                .filter((zuordnung: Zuordnung) => zuordnung.typ === OrganisationsTyp.Klasse)
+                .map((zuordnung: Zuordnung) => (zuordnung.sskName.length ? zuordnung.sskName : '---'))
+                .join(', ')
+            : '---';
+
+          const personalnummer: string = person.person.personalnummer ?? '---';
+
+          return {
+            ...person,
+            rollen: rollenZuordnungen,
+            administrationsebenen: administrationsebenen,
+            klassen: klassenZuordnungen,
+            person: { ...person.person, personalnummer: personalnummer },
+          };
+        });
       } catch (error: unknown) {
         this.errorCode = 'UNSPECIFIED_ERROR';
         if (isAxiosError(error)) {
@@ -231,6 +330,21 @@ export const usePersonStore: StoreDefinition<'personStore', PersonState, PersonG
           this.errorCode = error.response?.data.code || 'UNSPECIFIED_ERROR';
         }
         return await Promise.reject(this.errorCode);
+      } finally {
+        this.loading = false;
+      }
+    },
+    async getPersonenuebersichtById(personId: string): Promise<void> {
+      this.loading = true;
+      try {
+        const { data }: { data: DBiamPersonenuebersichtResponse } =
+          await personenuebersichtApi.dBiamPersonenuebersichtControllerFindPersonenuebersichtenByPerson(personId);
+        this.personenuebersicht = data;
+      } catch (error: unknown) {
+        this.errorCode = 'UNSPECIFIED_ERROR';
+        if (isAxiosError(error)) {
+          this.errorCode = error.response?.data.code || 'UNSPECIFIED_ERROR';
+        }
       } finally {
         this.loading = false;
       }
