+++ src/stores/PersonenkontextStore.ts
@@ -7,14 +7,11 @@ import {
   DbiamPersonenkontexteApiFactory,
   type DbiamPersonenkontexteApiInterface,
   type DBiamPersonenkontextResponse,
-  type DbiamPersonenuebersichtApiInterface,
   type DBiamPersonenuebersichtControllerFindPersonenuebersichten200Response,
-  type DBiamPersonenuebersichtResponse,
   type PersonenkontexteApiInterface,
   type SystemrechtResponse,
   PersonenkontextApiFactory,
   type PersonenkontextApiInterface,
-  DbiamPersonenuebersichtApiFactory,
   OrganisationsTyp,
   type DbiamUpdatePersonenkontexteBodyParams,
   type PersonenkontexteUpdateResponse,
@@ -28,6 +25,7 @@ import {
   RollenMerkmal,
 } from '../api-client/generated/api';
 import axiosApiInstance from '@/services/ApiService';
+import { usePersonStore, type PersonStore } from './PersonStore';
 
 const personenKontextApi: PersonenkontextApiInterface = PersonenkontextApiFactory(undefined, '', axiosApiInstance);
 const personenKontexteApi: PersonenkontexteApiInterface = PersonenkontexteApiFactory(undefined, '', axiosApiInstance);
@@ -36,12 +34,6 @@ const dbiamPersonenkontexteApi: DbiamPersonenkontexteApiInterface = DbiamPersone
   '',
   axiosApiInstance,
 );
-const personenuebersichtApi: DbiamPersonenuebersichtApiInterface = DbiamPersonenuebersichtApiFactory(
-  undefined,
-  '',
-  axiosApiInstance,
-);
-
 const personAdministrationApi: PersonAdministrationApiInterface = PersonAdministrationApiFactory(
   undefined,
   '',
@@ -66,28 +58,6 @@ export type Zuordnung = {
   merkmale: RollenMerkmal;
 };
 
-export type Uebersicht =
-  | {
-      personId: string;
-      vorname: string;
-      nachname: string;
-      benutzername: string;
-      lastModifiedZuordnungen: string | null;
-      zuordnungen: {
-        klasse?: string | undefined;
-        sskId: string;
-        rolleId: string;
-        sskName: string;
-        sskDstNr: string;
-        rolle: string;
-        administriertVon: string;
-        typ: OrganisationsTyp;
-        editable: boolean;
-        merkmale: RollenMerkmal;
-      }[];
-    }
-  | undefined;
-
 export type WorkflowFilter = {
   organisationId?: string;
   rolleId?: string;
@@ -102,7 +72,6 @@ type PersonenkontextState = {
   workflowStepResponse: PersonenkontextWorkflowResponse | null;
   filteredRollen: FindRollenResponse | null;
   filteredOrganisationen: FindSchulstrukturknotenResponse | null;
-  personenuebersicht: DBiamPersonenuebersichtResponse | null;
   createdPersonenkontextForOrganisation: DBiamPersonenkontextResponse | null;
   createdPersonenkontextForKlasse: DBiamPersonenkontextResponse | null;
   createdPersonWithKontext: DBiamPersonResponse | null;
@@ -122,8 +91,6 @@ type PersonenkontextActions = {
     personenkontext: DbiamPersonenkontextBodyParams,
     personenKontextTyp: PersonenKontextTyp,
   ) => Promise<DBiamPersonenkontextResponse>;
-  getPersonenuebersichtById: (personId: string) => Promise<void>;
-  getAllPersonenuebersichten: () => Promise<void>;
   createPersonWithKontext: (params: DbiamCreatePersonWithContextBodyParams) => Promise<PersonendatensatzResponse>;
 };
 
@@ -164,7 +131,6 @@ export const usePersonenkontextStore: StoreDefinition<
       updatedPersonenkontexte: null,
       filteredRollen: null,
       filteredOrganisationen: null,
-      personenuebersicht: null,
       createdPersonenkontextForOrganisation: null,
       createdPersonenkontextForKlasse: null,
       createdPersonWithKontext: null,
@@ -283,12 +249,13 @@ export const usePersonenkontextStore: StoreDefinition<
       }
     },
     async updatePersonenkontexte(combinedZuordnungen: Zuordnung[] | undefined, personId: string): Promise<void> {
+      const personStore: PersonStore = usePersonStore();
       this.loading = true;
       this.errorCode = '';
       try {
         const updateParams: DbiamUpdatePersonenkontexteBodyParams = {
-          lastModified: this.personenuebersicht?.lastModifiedZuordnungen ?? undefined,
-          count: this.personenuebersicht?.zuordnungen.length ?? 0,
+          lastModified: personStore.personenuebersicht?.lastModifiedZuordnungen ?? undefined,
+          count: personStore.personenuebersicht?.zuordnungen.length ?? 0,
           personenkontexte: combinedZuordnungen?.map((zuordnung: Zuordnung) => ({
             personId: personId,
             organisationId: zuordnung.sskId,
@@ -302,36 +269,6 @@ export const usePersonenkontextStore: StoreDefinition<
         this.errorCode = 'UNSPECIFIED_ERROR';
         if (isAxiosError(error)) {
           this.errorCode = error.response?.data.i18nKey || 'PERSONENKONTEXTE_UPDATE_ERROR';
-        }
-      } finally {
-        this.loading = false;
-      }
-    },
-    async getPersonenuebersichtById(personId: string): Promise<void> {
-      this.loading = true;
-      try {
-        const { data }: { data: DBiamPersonenuebersichtResponse } =
-          await personenuebersichtApi.dBiamPersonenuebersichtControllerFindPersonenuebersichtenByPerson(personId);
-        this.personenuebersicht = data;
-      } catch (error: unknown) {
-        this.errorCode = 'UNSPECIFIED_ERROR';
-        if (isAxiosError(error)) {
-          this.errorCode = error.response?.data.code || 'UNSPECIFIED_ERROR';
-        }
-      } finally {
-        this.loading = false;
-      }
-    },
-    async getAllPersonenuebersichten(): Promise<void> {
-      this.loading = true;
-      try {
-        const { data }: { data: DBiamPersonenuebersichtControllerFindPersonenuebersichten200Response } =
-          await personenuebersichtApi.dBiamPersonenuebersichtControllerFindPersonenuebersichten();
-        this.allUebersichten = data;
-      } catch (error: unknown) {
-        this.errorCode = 'UNSPECIFIED_ERROR';
-        if (isAxiosError(error)) {
-          this.errorCode = error.response?.data.code || 'UNSPECIFIED_ERROR';
         }
       } finally {
         this.loading = false;
