+++ src/views/ProfileView.vue
@@ -10,9 +10,10 @@
   };
 
   import { usePersonInfoStore, type PersonInfoStore, type PersonInfoResponse } from '@/stores/PersonInfoStore';
-  import { usePersonenkontextStore, type Zuordnung, type PersonenkontextStore } from '@/stores/PersonenkontextStore';
+  import { type Zuordnung } from '@/stores/PersonenkontextStore';
   import { OrganisationsTyp } from '@/stores/OrganisationStore';
   import { type RouteLocationNormalizedLoaded, type Router, useRoute, useRouter } from 'vue-router';
+  import { usePersonStore, type PersonStore } from '@/stores/PersonStore';
 
   const route: RouteLocationNormalizedLoaded = useRoute();
   const router: Router = useRouter();
@@ -26,7 +27,7 @@
   };
 
   let personInfoStore: PersonInfoStore = usePersonInfoStore();
-  let personenkontextStore: PersonenkontextStore = usePersonenkontextStore();
+  let personStore: PersonStore = usePersonStore();
   const personalData: Ref = ref<LabelValue[]>([]);
   const schulDaten: Ref = ref<SchulDaten[]>([]);
 
@@ -133,9 +134,9 @@
 
   async function initializeStores(): Promise<void> {
     personInfoStore = usePersonInfoStore();
-    personenkontextStore = usePersonenkontextStore();
+    personStore = usePersonStore();
     await personInfoStore.initPersonInfo();
-    await personenkontextStore.getPersonenuebersichtById(personInfoStore.personInfo?.person.id ?? '');
+    await personStore.getPersonenuebersichtById(personInfoStore.personInfo?.person.id ?? '');
   }
 
   function setupPersonalData(): void {
@@ -159,8 +160,8 @@
   }
 
   function setupSchuleData(): void {
-    if (!personenkontextStore.personenuebersicht) return;
-    const personenZuordnungen: Zuordnung[] = personenkontextStore.personenuebersicht.zuordnungen;
+    if (!personStore.personenuebersicht) return;
+    const personenZuordnungen: Zuordnung[] = personStore.personenuebersicht.zuordnungen;
     const groupedZuordnungen: Map<string, Zuordnung[]> = groupZuordnungen(
       personenZuordnungen.map(
         (z: Zuordnung) =>
